+++ app/helpers/rest_url.rb
